% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificialLongiData.R
\docType{data}
\name{artificialLongiData}
\alias{artificialLongiData}
\title{Artificial longitudinal data.}
\format{A reshaped data frame with eight continuous variables: \code{x1}, \code{x2},
\code{x3}, \code{x4}, \code{x5}, \code{x6}, \code{x7}, and \code{x8}.}
\usage{
artificialLongiData
}
\description{
A data set of 200 intances, that is generated from a network consisting of four variables,
and three time slices.
}
\keyword{datasets}

