% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossdataV6.R
\docType{data}
\name{crossdata6V}
\alias{crossdata6V}
\title{Artificial cross-sectional data.}
\format{A data frame with six continuous variables: \code{x1,...,x6}.}
\usage{
crossdata6V
}
\description{
A data set of 500 intances, generated from a network consisting
of six continuous variables.
}
\keyword{datasets}

