% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longiData4V3T.R
\docType{data}
\name{longiData4V3T}
\alias{longiData4V3T}
\title{Artificial longitudinal data.}
\format{A data frame with twelve continuous variables:
\code{x1,...,x4} are for time slice \code{t_0},
\code{x5,...,x8} are for time slice \code{t_1}, and
\code{x9,...,x12} are for time slice \code{t_2}}
\usage{
longiData4V3T
}
\description{
A data set of 400 intances, that is generated from a network consisting of
four continuous variables and three time slices \code{t_0,...,t_2}.
}
\keyword{datasets}

